address review: shared bootstrap helper + GroupDeleted docs
- Extract `bootstrap_namespace_with_admin` test helper (meta + member +
  identity bootstrap) and use it in the two GroupDeleted apply tests,
  removing copy-paste setup (review nit).
- Update architecture/local-governance.html GroupDeleted entry: each peer
  re-enumerates and deletes only the locally-recomputed subtree; payload
  cascade ids feed the subset/divergence check only and are never trusted
  as the delete set.
