+++ architecture/local-governance.html
@@ -441,7 +441,7 @@
     <p style="margin-bottom:8px;font-size:.82rem;">Visible to all namespace members. Used for structural changes and key distribution:</p>
     <div class="fr"><div class="fp">NamespaceCreated { founder }</div><div class="fd"><strong>The namespace-root genesis op (#2474)</strong> — the first op in a namespace DAG: <strong>no parents</strong>, nonce 1 (the head defaults <code>next_nonce</code> to 1 when absent), signed by the namespace key (= the <code>founder</code>). It is <strong>self-authorizing</strong>: it establishes the founding admin and does not require a pre-existing admin. On apply it writes the root meta with <code>admin_identity = owner_identity = founder</code>, an <code>Admin</code> member row for the founder, and the default <span class="code">CAN_JOIN_OPEN_SUBGROUPS</span> capability. Emitted by <span class="code">create_group.rs</span> on root creation (<code>parent_group_id == None</code>), which previously emitted no governance op at all. <strong>Anti-hijack guards:</strong> a <code>NamespaceCreated</code> whose <code>signer != founder</code> is rejected (<code>NamespaceCreatedRejection::SignerNotFounder</code>); a <code>NamespaceCreated</code> that carries parents on a bare namespace is rejected (<code>NamespaceCreatedRejection::NotGenesis</code>) and does not advance the DAG head, so a subsequent parentless genesis can still apply. A parented arrival on an already-established namespace is a no-op <code>Ok</code>. <strong>Wire-breaking Borsh addition:</strong> appended at the end of <code>RootOp</code> so existing discriminants do not renumber, but consumers pinning this crate must coordinate a reset/core-rev bump.</div></div>
     <div class="fr"><div class="fp">GroupCreated</div><div class="fd">Atomically creates a new group AND nests it under <code>parent_id</code> (strict-tree invariant — no orphan creation path). Carries a <span class="code">restricted</span> boolean field (<strong>wire-breaking Borsh change</strong>); apply handler writes subgroup visibility immediately when <code>restricted: false</code>, and triggers a buffered-op retry if the node already holds a key for the new group.</div></div>
-    <div class="fr"><div class="fp">GroupDeleted</div><div class="fd">Cascade-deletes a group plus its entire subtree and all contained contexts in one op. Payload carries <code>cascade_group_ids</code> + <code>cascade_context_ids</code> which every peer re-enumerates and verifies for deterministic application.</div></div>
+    <div class="fr"><div class="fp">GroupDeleted</div><div class="fd">Cascade-deletes a group plus its entire subtree and all contained contexts in one op. Authorization is checked only against <code>root_group_id</code> (the subgroup owner, a namespace-root admin, or a holder of <code>CAN_DELETE_SUBGROUP</code>). The op also carries <code>cascade_group_ids</code> + <code>cascade_context_ids</code>, but each peer <strong>re-enumerates the subtree locally and deletes only that locally-recomputed set</strong> &mdash; the payload ids are never trusted as the delete set. They feed a subset/divergence check only (local subtree must be covered by the payload, so no peer under-deletes relative to what the network agreed on); payload ids outside the local subtree are ignored, not deleted. This prevents a leaf-subgroup owner from listing arbitrary &mdash; even cross-namespace &mdash; ids and having every peer delete them.</div></div>
     <div class="fr"><div class="fp">GroupReparented</div><div class="fd">Atomic edge swap: moves a child from its current parent to a new parent within the same namespace. Replaces the previous <code>GroupNested</code> + <code>GroupUnnested</code> pair; orphan state is structurally impossible.</div></div>
     <div class="fr"><div class="fp">MemberJoined</div><div class="fd">Legacy cleartext join notification (joiner doesn&rsquo;t have group key yet). Carries an admin-signed <span class="code">SignedGroupOpenInvitation</span>; the invitation&rsquo;s <code>expiration_timestamp</code> must be <code>0</code> (no-expiry sentinel) for this variant — non-zero expiration requires <code>MemberJoinedAt</code> instead.</div></div>
     <div class="fr"><div class="fp">MemberJoinedAt</div><div class="fd">Invitation-based join carrying the joiner&rsquo;s claimed redemption timestamp (<code>joined_at</code>, unix seconds, covered by the joiner&rsquo;s signature). Apply deterministically rejects the op when <code>joined_at &gt; expiration_timestamp</code>, enforcing expiry on every node without relying on a local wall clock. Use this variant whenever the invitation has a non-zero <code>expiration_timestamp</code>.</div></div>
